Revert r26584: "Upgrade to Perst 4.05"
This is causing more problem then it supposed to fix.

git-svn-id: file:///home/toad/git-migration/temprepository/trunk/plugins/XMLSpider@26646 67a373e5-eb02-0410-a15c-ee090a768436

+++ src/plugins/XMLSpider/org/garret/perst/impl/ReplicationMasterStorageImpl.java
@@ -1,7 +1,6 @@
 package plugins.XMLSpider.org.garret.perst.impl;
 
-import plugins.XMLSpider.org.garret.perst.IFile;
-import plugins.XMLSpider.org.garret.perst.ReplicationMasterStorage;
+import plugins.XMLSpider.org.garret.perst.*;
 
 
 public class ReplicationMasterStorageImpl extends StorageImpl implements ReplicationMasterStorage
@@ -12,7 +11,7 @@ public class ReplicationMasterStorageImpl extends StorageImpl implements Replica
         this.asyncBufSize = asyncBufSize;
     }
     
-    public void open(IFile file, long pagePoolSize) {
+    public void open(IFile file, int pagePoolSize) {
         super.open(asyncBufSize != 0 
                    ? (ReplicationMasterFile)new AsyncReplicationMasterFile(this, file, asyncBufSize)
                    : new ReplicationMasterFile(this, file),
