+++ src/plugins/XMLSpider/org/garret/perst/impl/ReplicationStaticSlaveStorageImpl.java
@@ -1,11 +1,9 @@
 package plugins.XMLSpider.org.garret.perst.impl;
 
-import java.io.IOException;
-import java.net.ServerSocket;
-import java.net.Socket;
+import java.io.*;
+import java.net.*;
 
-import plugins.XMLSpider.org.garret.perst.IFile;
-import plugins.XMLSpider.org.garret.perst.StorageError;
+import plugins.XMLSpider.org.garret.perst.*;
 
 
 public class ReplicationStaticSlaveStorageImpl extends ReplicationSlaveStorageImpl
@@ -14,7 +12,7 @@ public class ReplicationStaticSlaveStorageImpl extends ReplicationSlaveStorageIm
         this.port = port;
     }
 
-    public void open(IFile file, long pagePoolSize) {
+    public void open(IFile file, int pagePoolSize) {
         try { 
             acceptor = new ServerSocket(port);
         } catch (IOException x) {
